requirements.txt
sqlalchemy>=2.0
psycopg2>=2.9
requests>=2.31
praw>=7.7
newsapi-python>=0.2.7
apscheduler>=3.10
